+++ brainsets/brainsets/runner.py
@@ -140,10 +140,8 @@ def run():
     if pipeline_cls.parser is not None:
         pipeline_args = pipeline_cls.parser.parse_args(remaining_args)
 
-    # Set raw and processed dir
     raw_dir = args.raw_dir / pipeline_cls.brainset_id
-    processed_dir = args.processed_dir / pipeline_cls.brainset_id
-
+    raw_dir.mkdir(parents=True, exist_ok=True)
     manifest = pipeline_cls.get_manifest(
         raw_dir=raw_dir,
         args=pipeline_args,
@@ -155,6 +153,8 @@ def run():
             print(manifest)
         sys.exit(0)
 
+    processed_dir = args.processed_dir / pipeline_cls.brainset_id
+    processed_dir.mkdir(parents=True, exist_ok=True)
     if args.single is None:
         # Parallel run