+++ brainsets/brainsets_pipelines/odoherty_sabes_nonhuman_2017/pipeline.py
@@ -39,8 +39,6 @@ class Pipeline(BrainsetPipeline):
 
     @classmethod
     def get_manifest(cls, raw_dir, args) -> pd.DataFrame:
-        raw_dir.mkdir(exist_ok=True, parents=True)
-
         # list files to download
         r_fd = os.popen(f"zenodo_get -w - 3854034")
         manifest_out = r_fd.read()
@@ -90,7 +88,6 @@ class Pipeline(BrainsetPipeline):
 
     def download(self, manifest_item):
         self.update_status("DOWNLOADING")
-        self.raw_dir.mkdir(exist_ok=True, parents=True)
         fpath = self.raw_dir / Path(manifest_item.file)
 
         if (
@@ -120,7 +117,6 @@ class Pipeline(BrainsetPipeline):
         return fpath
 
     def process(self, fpath):
-        self.processed_dir.mkdir(exist_ok=True, parents=True)
 
         brainset_description = BrainsetDescription(
             id="odoherty_sabes_nonhuman_2017",